Only listen if message doesn't come from a bot

+++ src/main/kotlin/me/obsilabor/noriskclientbot/discord/listener/MessageListener.kt
@@ -34,6 +34,9 @@ class MessageListener : Listener {
 
     override suspend fun register(client: Kord) {
         client.on<MessageCreateEvent> {
+            if(this.message.author?.isBot == true) {
+                return@on
+            }
             if(this.message.containsInvite()) {
                 this.message.delete()
             }
